Update the setting.vue
This is the web I used for the code of logout: https://stackoverflow.com/questions/50181321/laravel-logout-route-in-vue-component. So basically “Home/My profile/ Logout”. The home and my profile need the link to that. I imported the vue files in the script just in case. Need to add revise the link of home and my profile. Check for the font. Thx